agent/requirements.txt
strands-agents[anthropic]
python-dotenv>=1.0.0
